# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatMessageOut": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "role": {
            "$ref": "#/components/schemas/ChatRole"
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "image_url",
          "created_at"
        ],
        "title": "ChatMessageOut",
        "type": "object"
      },
      "ChatRequest": {
        "properties": {
          "image_key": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Key"
          },
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatRole": {
        "enum": [
          "user",
          "assistant"
        ],
        "title": "ChatRole",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PendingActionOut": {
        "properties": {
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action_type",
          "summary",
          "payload",
          "status",
          "created_at"
        ],
        "title": "PendingActionOut",
        "type": "object"
      },
      "PendingActionPatch": {
        "description": "Onaydan önce öneriyi düzenlemek için.\n\nBölüm 4.1'in şartı: AI'nın önerdiği program tam ekran bir \"gözden geçir\"\nekranında satır satır düzenlenebilmeli, sonra onaylanmalı.\n\nPayload'ı istemcinin yeniden yazması güvenlik açığı DEĞİL: istemci zaten\nkullanıcının kendisi ve payload onay anında Pydantic ile yeniden doğrulanıyor.\nAsıl güvence orada — burada ne yazılırsa yazılsın, şemaya uymayan hiçbir şey\nveritabanına geçemez.",
        "properties": {
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "payload"
        ],
        "title": "PendingActionPatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Progresif overload merkezli antrenman, beslenme ve sağlık takip sistemi.",
    "title": "overload API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chat/messages": {
      "get": {
        "operationId": "list_messages_chat_messages_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatMessageOut"
                  },
                  "title": "Response List Messages Chat Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Messages",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/pending-actions": {
      "get": {
        "operationId": "list_pending_chat_pending_actions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PendingActionOut"
                  },
                  "title": "Response List Pending Chat Pending Actions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Pending",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/pending-actions/{action_id}": {
      "get": {
        "description": "Tek bir öneriyi getirir — gözden geçirme ekranı bunu okuyor.",
        "operationId": "get_pending_action_chat_pending_actions__action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Pending Action",
        "tags": [
          "chat"
        ]
      },
      "patch": {
        "description": "Onay bekleyen öneriyi düzenler. Yalnızca `pending` durumdayken.",
        "operationId": "update_pending_action_chat_pending_actions__action_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PendingActionPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Pending Action",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/pending-actions/{action_id}/approve": {
      "post": {
        "description": "Onaylanan aksiyonu uygular.\n\n`executors.apply_pending_action` payload'ı Pydantic ile YENİDEN doğrular —\npayload modelin yazdığı veri olduğu için güvenilmez kabul edilir.",
        "operationId": "approve_action_chat_pending_actions__action_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Approve Action",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/pending-actions/{action_id}/reject": {
      "post": {
        "operationId": "reject_action_chat_pending_actions__action_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Reject Action",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/stream": {
      "post": {
        "description": "Bir sohbet turunu Server-Sent Events olarak akıtır.\n\nAkış SSE çünkü tek yönlü: sunucudan istemciye token akışı. WebSocket çift yönlü\nkanal kurup yeniden bağlanma/kalp atışı yönetimi getirirdi; burada karşılığı yok.",
        "operationId": "stream_chat_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Stream Chat",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}